fix: Json response

+++ backend/com.desafiobasis/src/main/java/com/desafiobasis/com/desafiobasis/Repository/EmpresaRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
 
-    Empresa findById(long id);
+    Empresa findById(int id);
 
-    Empresa deleteById(long id);
+    Empresa deleteById(int id);
 
 
 }
